Split tryLogCurrentException(LogFrequencyLimiterImpl) into separate module
To avoid extra dependencies on DateLUT for examples.

CI: https://s3.amazonaws.com/clickhouse-test-reports/PRs/97171/d70cdd8f26eb92b78869fb0c819daa4d195a6dbb/build_amd_compat/job.log
